[TA3684] fix(rebuild): create new internal snap for every rebuild (#137)
[TA3684] fix(rebuild): create new internal snap for every rebuild

For every rebuild, a new internal snapshot need to be created on 
the helping replica. But, currently, if a new internal snapshots 
exists, it will use it in the rebuild process.

This change creates new internal snapshot on the helping replica for every rebuild request.

Also, currently, on any error of opening/holding the rebuild snapshot or clone, 
rebuild snapshot and clone gets deleted. This change comments out that code.

Signed-off-by: Vishnu Itta <vitta@mayadata.io>